Review comment: move layer functions from geometry to napari_utils. Add docstring to point_in_layer_bbox
