+++ src/napari_threedee/utils/geometry.py
@@ -87,40 +87,3 @@ def point_in_bounding_box(point: np.ndarray, bounding_box: np.ndarray) -> bool:
     if np.all(point >= bounding_box[0]) and np.all(point <= bounding_box[1]):
         return True
     return False
-
-
-def point_in_layer_bounding_box(point, layer):
-    dims_displayed = get_dims_displayed(layer)
-    bbox = layer._display_bounding_box(dims_displayed).T
-    if np.any(point < bbox[0]) or np.any(point > bbox[1]):
-        return False
-    else:
-        return True
-
-def clamp_point_to_layer_bounding_box(point: np.ndarray, layer):
-    """Ensure that a point is inside of the bounding box of a given layer. 
-    If the point has a coordinate outside of the bounding box, the value 
-    is clipped to the max extent of the bounding box.
-
-    Parameters
-    ----------
-    point : np.ndarray
-        n-dimensional point as an (n,) ndarray. Multiple points can
-        be passed as an (n, D) array.
-    layer : napari.layers.Layer
-        napari layer to get the bounding box from
-
-    Returns
-    -------
-    clamped_point : np.ndarray
-        `point` clamped to the limits of the layer bounding box
-
-    Notes
-    -----
-    This function is derived from the napari function:
-        `napari.utils.geometry.clamp_point_to_bounding_box`
-    """
-    dims_displayed = get_dims_displayed(layer)
-    bbox = layer._display_bounding_box(dims_displayed)
-    clamped_point = np.clip(point, bbox[:, 0], bbox[:, 1] - 1)
-    return clamped_point
